Bind backfill+monitor+LRU to percent of runs to evict
This change follows up from #935, which created EvictLRU(percent) in the LRU. This change passes a percentage of runs to evict to FillIndex(), which is responsible for setting up a monitor for the given index. The percentage is passed from the Monitor through the Index to the LRU when performing a synchronized operation to evict a collection of LRU runs to reducem memory pressure.